feat(primitives): move virtual to primitives
