+++ packages/components/src/checkbox/stories/checkbox-integrations.stories.tsx
@@ -8,6 +8,7 @@ import { Stack } from '../../stack'
 import { Button } from '../../button'
 import { Text } from '../../text'
 import { Field, FieldDescription } from '../../field'
+import { Virtual, VirtualItem } from '@vtex/shoreline-primitives'
 
 export default {
   title: 'shoreline-components/checkbox/integrations',
@@ -140,6 +141,67 @@ export function ReactWindow() {
           </div>
         )}
       </FixedSizeList>
+    </Stack>
+  )
+}
+
+export function VirtualPrimitive() {
+  const numberOfItems = useMemo(() => 50000, [])
+
+  const items = new Array(numberOfItems).fill(true)
+
+  const trueArray = useMemo(
+    () => new Array(numberOfItems).fill(true),
+    [numberOfItems]
+  )
+
+  const falseArray = useMemo(
+    () => new Array(numberOfItems).fill(false),
+    [numberOfItems]
+  )
+
+  const [checked, setChecked] = useState<boolean[]>(falseArray)
+
+  const someChecked = checked.some((i) => i)
+  const allChecked = checked.every((i) => i)
+
+  return (
+    <Stack>
+      <Text> Number of Checkboxes: {numberOfItems}</Text>
+
+      <Checkbox
+        indeterminate={someChecked && !allChecked}
+        checked={allChecked}
+        onChange={() => {
+          if (allChecked) {
+            setChecked(falseArray)
+          } else {
+            setChecked(trueArray)
+          }
+        }}
+      >
+        Root
+      </Checkbox>
+      <Virtual count={items.length} dynamic>
+        <VirtualItem asChild>
+          {({ index }) => (
+            <Checkbox
+              checked={checked[index]}
+              onChange={() => {
+                setChecked((prev) => {
+                  const res = [...prev]
+
+                  res[index] = !res[index]
+
+                  return res
+                })
+              }}
+            >
+              Item {index}
+            </Checkbox>
+          )}
+        </VirtualItem>
+      </Virtual>
     </Stack>
   )
 }
