+++ packages/components/src/index.ts
@@ -10,7 +10,6 @@ export * from './skeleton'
 export * from './tab'
 export * from './tooltip'
 export * from './filter'
-export * from './virtual'
 export * from './link'
 export * from './pagination'
 export * from './modal'
