--- packages/components/src/virtual/stories/styles.css
@@ -1,6 +1,0 @@
-.row {
-  background: white;
-  &[data-odd='true'] {
-    background: #cecece;
-  }
-}
