+++ packages/primitives/README.md
@@ -1,0 +1,21 @@
+# Shoreline Primitives
+
+Primitives are headless componenents that are use as base for Shoreline components. If you are using the `@vtex/shoreline-components`, you don't need to install this package.
+
+## Instalation
+
+```sh
+pnpm add @vtex/shoreline-primitives
+```
+
+```sh
+bun add @vtex/shoreline-primitives
+```
+
+```sh
+npm i @vtex/shoreline-primitives
+```
+
+```sh
+yarn add @vtex/shoreline-primitives
+```
