+++ packages/primitives/src/index.ts
@@ -6,3 +6,4 @@ export * from './compose'
 export * from './accessible-icon'
 export * from './visually-hidden'
 export * from './locale'
+export * from './virtual'
