+++ packages/primitives/src/virtual/index.ts
@@ -1,4 +1,4 @@
 export * from './virtual'
 export * from './virtual-item'
 export * from './virtual-container'
-export * from './useVirtualizerModel'
+export * from './use-virtualizer-model'
