+++ packages/primitives/src/virtual/virtual-container.tsx
@@ -1,8 +1,9 @@
 import type { ComponentPropsWithoutRef, ReactNode } from 'react'
 import React, { Fragment, cloneElement } from 'react'
 import { forwardRef } from '@vtex/shoreline-utils'
-import type { UseVirtualizerModelReturn } from './useVirtualizerModel'
-import { Compose } from '@vtex/shoreline-primitives'
+import type { UseVirtualizerModelReturn } from './use-virtualizer-model'
+
+import { Compose } from '../compose'
 
 export const VirtualContainer = forwardRef<
   HTMLDivElement,
